+++ comment-updater-headless/src/main/kotlin/PluginRunner.kt
@@ -4,21 +4,17 @@ import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.application.ApplicationStarter
 import com.intellij.openapi.components.ServiceManager
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.roots.ProjectRootManager
 import com.intellij.openapi.vcs.ProjectLevelVcsManager
 import com.intellij.openapi.vcs.VcsException
 import com.intellij.openapi.vcs.changes.Change
 import com.intellij.openapi.vcs.impl.ProjectLevelVcsManagerImpl
-import com.intellij.openapi.vfs.VfsUtilCore
 import com.intellij.psi.PsiFile
 import com.intellij.psi.PsiFileFactory
-import com.intellij.psi.PsiManager
 import com.intellij.psi.PsiMethod
 import com.intellij.psi.util.PsiTreeUtil
 import git4idea.GitCommit
 import git4idea.GitVcs
 import git4idea.repo.GitRepositoryManager
-import gr.uom.java.xmi.diff.RenameOperationRefactoring
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
@@ -128,26 +124,10 @@ class CodeCommentExtractor {
 
 
     fun collectProjectExamples(projectPath: String) {
-        val latch = CountDownLatch(1)
-
         val project = ProjectUtil.openOrImport(projectPath, null, true)
         if( project == null) {
             log(LogLevel.WARN, "Can't open project $projectPath")
             return
-        }
-        val projectPsiFiles = mutableListOf<PsiFile>()
-
-        ApplicationManager.getApplication().runReadAction {
-            ProjectRootManager.getInstance(project).contentRoots.mapNotNull { root ->
-                VfsUtilCore.iterateChildrenRecursively(root, null) { virtualFile ->
-                    if (virtualFile.extension != "java" || virtualFile.canonicalPath == null) {
-                        return@iterateChildrenRecursively true
-                    }
-                    val psi =
-                        PsiManager.getInstance(project).findFile(virtualFile) ?: return@iterateChildrenRecursively true
-                    projectPsiFiles.add(psi)
-                }
-            }
         }
 
         val vcsManager = ServiceManager.getService(
@@ -163,6 +143,7 @@ class CodeCommentExtractor {
         // Checkout https://intellij-support.jetbrains.com/hc/en-us/community/posts/206105769-Get-project-git-repositories-in-a-project-component
         // To understand why we should call runAfterInitialization and create this runnable
 
+        val latch = CountDownLatch(1)
         val runnable = {
             try {
                 val gitRoots = vcsManager.getRootsUnderVcs(GitVcs.getInstance(project))
@@ -170,11 +151,11 @@ class CodeCommentExtractor {
                     val repo = gitRepoManager.getRepositoryForRoot(root) ?: continue
                     runBlocking {
                         val commits = repo.walkAll()
-                        val numOfCommits = commits.size
+                        statsHandler.numberOfCommits = commits.size
 
                         commits.map { commit ->
                             async(Dispatchers.Default) {
-                                processCommit(commit, numOfCommits, project)
+                                processCommit(commit, project)
                             }
                         }.awaitAll()
                     }
@@ -193,18 +174,18 @@ class CodeCommentExtractor {
         latch.await()
     }
 
+
     fun processCommit(
         commit: GitCommit,
-        numOfCommits: Int,
         project: Project
     ) {
         try {
-            getChanges(commit, ".java").forEach { change ->
+            commit.filterChanges(".java").forEach { change ->
                 val fileName = change.afterRevision?.file?.name ?: ""
                 log(
                     LogLevel.INFO,
                     "Commit: ${commit.id.toShortString()} num ~ ${statsHandler.processedCommits.get()}" +
-                            "/$numOfCommits File changed: $fileName",
+                            "/${statsHandler.numberOfCommits} File changed: $fileName",
                     logThread = true
                 )
 
@@ -232,7 +213,7 @@ class CodeCommentExtractor {
         val methodsRefactorings = RefactoringExtractor.methodsToRefactoringTypes(refactorings)
 
         val changedMethods = try {
-            extractChangedMethods(project, change, refactorings)
+            ProjectMethodExtractor.extractChangedMethods(project, change, refactorings)
         } catch (e: VcsException) {
             log(LogLevel.WARN, "Unexpected VCS exception", logThread = true)
             e.printStackTrace()
@@ -295,112 +276,8 @@ class CodeCommentExtractor {
 
 
 
-    /**
-     * @return: Map from old method name to new method name
-     */
-    fun extractNameChanges(refactorings: List<Refactoring>): HashMap<String, String> {
-        return hashMapOf(*refactorings.filter {
-            it.refactoringType == RefactoringType.RENAME_METHOD
-        }.map {
-            val renameRefactoring = (it as RenameOperationRefactoring)
-            (renameRefactoring.renamedOperation.className + "." + renameRefactoring.renamedOperation.name
-                    to
-                    renameRefactoring.originalOperation.className + "." + renameRefactoring.originalOperation.name)
-        }.toTypedArray())
-    }
-
-    /**
-     * @return: List of pairs oldMethod to newMethod
-     */
-    fun extractChangedMethods(
-        project: Project,
-        change: Change,
-        refactorings: List<Refactoring>
-    ): MutableList<Pair<PsiMethod, PsiMethod>> {
-        val before = change.beforeRevision?.content ?: return mutableListOf()
-        val after = change.afterRevision?.content ?: return mutableListOf()
-
-        val renameMapping = extractNameChanges(refactorings)
-
-        val changedMethodPairs = mutableListOf<Pair<PsiMethod, PsiMethod>>()
-
-        val oldNamesToMethods = extractNamesToMethods(project, before)
-
-        val newMethods = extractMethodsWithNames(project, after)
-
-        newMethods.forEach { (afterName, newMethod) ->
-            val beforeName = if (renameMapping.containsKey(afterName)) {
-                renameMapping[afterName]
-            } else {
-                afterName
-            }
-            oldNamesToMethods.get(beforeName)?.let { oldMethod ->
-                changedMethodPairs.add(oldMethod to newMethod)
-            }
 
 
-        }
-        return changedMethodPairs
-    }
-
-    fun extractMethodsWithNames(project: Project, content: String): List<Pair<String, PsiMethod>> {
-        lateinit var psiFile: PsiFile
-        lateinit var methodsWithNames: List<Pair<String, PsiMethod>>
-
-        ApplicationManager.getApplication().runReadAction {
-            psiFile = PsiFileFactory.getInstance(project).createFileFromText(
-                "extractMethodsWithNamesFile",
-                JavaFileType.INSTANCE,
-                content
-            )
-
-            methodsWithNames = PsiTreeUtil.findChildrenOfType(psiFile, PsiMethod::class.java).filter {
-                it.docComment != null
-            }.map {
-                ((it.containingClass?.qualifiedName ?: "") + "." + it.name) to it
-            }
-
-        }
-
-        return methodsWithNames
-    }
-
-    fun extractNamesToMethods(project: Project, content: String): HashMap<String, PsiMethod> {
-        lateinit var psiFile: PsiFile
-        lateinit var methods: List<PsiMethod>
-        lateinit var namesToMethods: HashMap<String, PsiMethod>
-
-        ApplicationManager.getApplication().runReadAction {
-            psiFile = PsiFileFactory.getInstance(project).createFileFromText(
-                "extractNamesToMethodsFile",
-                JavaFileType.INSTANCE,
-                content
-            )
-
-            methods = PsiTreeUtil.findChildrenOfType(psiFile, PsiMethod::class.java).filter {
-                it.docComment != null
-            }
-
-            namesToMethods = hashMapOf(*methods.map {
-                ((it.containingClass?.qualifiedName ?: "") + "." + it.name) to it
-            }.toTypedArray())
-        }
-
-        return namesToMethods
-    }
-
-    fun getChanges(commit: GitCommit, fileSuffix: String): List<Change> {
-        return commit.changes
-            .filter {
-                it.afterRevision != null
-                        &&
-                        it.beforeRevision != null
-            }
-            .filter {
-                // not null and true
-                it.virtualFile?.name?.endsWith(fileSuffix) == true
-            }
-    }
 
 
 }
